Make ManualInputHandler private
