+++ osu.Framework/Testing/Input/ManualInputManager.cs
@@ -3,8 +3,11 @@
 
 using osu.Framework.Graphics;
 using osu.Framework.Input;
+using osu.Framework.Input.Handlers;
+using osu.Framework.Platform;
 using OpenTK;
 using OpenTK.Input;
+using MouseState = osu.Framework.Input.MouseState;
 
 namespace osu.Framework.Testing.Input
 {
@@ -34,6 +37,34 @@ namespace osu.Framework.Testing.Input
         {
             UseParentState = false;
             handler.Click(button);
+        }
+
+        private class ManualInputHandler : InputHandler
+        {
+            private Vector2 lastMousePosition;
+
+            public void MoveMouseTo(Vector2 position)
+            {
+                PendingStates.Enqueue(new InputState { Mouse = new MouseState { Position = position } });
+                lastMousePosition = position;
+            }
+
+            public void Click(MouseButton button)
+            {
+                var mouseState = new MouseState { Position = lastMousePosition };
+                mouseState.SetPressed(button, true);
+
+                PendingStates.Enqueue(new InputState { Mouse = mouseState });
+
+                mouseState = (MouseState)mouseState.Clone();
+                mouseState.SetPressed(button, false);
+
+                PendingStates.Enqueue(new InputState { Mouse = mouseState });
+            }
+
+            public override bool Initialize(GameHost host) => true;
+            public override bool IsActive => true;
+            public override int Priority => 0;
         }
     }
 }
