+++ osu.Framework/osu.Framework.csproj
@@ -456,7 +456,6 @@
     <Compile Include="Testing\DynamicClassCompiler.cs" />
     <Compile Include="Testing\GridTestCase.cs" />
     <Compile Include="Testing\IDynamicallyCompile.cs" />
-    <Compile Include="Testing\Input\ManualInputHandler.cs" />
     <Compile Include="Testing\Input\ManualInputManager.cs" />
     <Compile Include="Testing\TestBrowser.cs" />
     <Compile Include="Testing\TestBrowserConfig.cs" />
